move REST-specific errors to REST API
